delivery - hide/unhide items

+++ src/components/Cart.js
@@ -24,7 +24,7 @@ export class Cart {
     this._render();
     this._init();
 
-    this.client = new Client(ORDER.client);
+    this.client = new Client(ORDER.client, this.availableProducts, this.unavailableProducts);
 
   }
 
@@ -133,25 +133,40 @@ export class Cart {
   _init() {
     this.orderContainer.addEventListener('click', e => {
 
+      // Увеличить количество товара в корзине
       if (e.target.classList.contains(ADD_CLASSNAME)) {
+
+
         const id = +e.target.dataset['id'];
         let item = this.getItem(id);
-        item.changeQuantity(1);
-        // this.updateCart();
+
+        if (item.productQuantity < item.productInStock) {
+          item.changeQuantity(1);
+          // this.updateCart();
+        }
+
       }
 
+      // Уменьшить количество товара в корзине
       if (e.target.classList.contains(DELETE_ONE_CLASSNAME)) {
         const id = +e.target.dataset['id'];
         let item = this.getItem(id);
-        this.deleteItem(item);
-        // this.updateCart();
+
+        if(item.productQuantity > 1) {
+          this.deleteItem(item);
+          // this.updateCart();
+        }
       }
 
+      // Удалить товар
       if (e.target.classList.contains(DELETE_CLASSNAME)) {
         const id = +e.target.dataset['id'];
         let item = this.getItem(id);
         this.deleteAll(item);
         // this.updateCart();
+
+        let elements = document.querySelectorAll(`.dlv[data-id="${id}"]`);
+        elements.forEach(elem => elem.remove());
       }
 
       if (e.target.classList.contains(DELETE_UNAVAILABLE_CLASSNAME)) {
@@ -159,28 +174,63 @@ export class Cart {
         let item = this.getItemUnavailable(id);
         this.deleteUnavailable(item);
       }
+
 //---------------
+      // СЛУШАТЕЛЬ СОБЫТИЯ НА ОБЩЕМ ЧЕКБОКСЕ - переключает отображение товаров в разделе доставки
       if (e.target.classList.contains(ICON_FAVORITE_SELECTOR)) {
         e.target.classList.toggle('cart__favorite_colored');
       }
 
       if (e.target.classList.contains('checkbox__input-all')) {
+        let deliveryDatesContainer = document.querySelector('.delivery__dates');
+
         let checkboxes = document.getElementsByClassName('checkbox__input');
         if (document.querySelector('.checkbox__input-all').checked) {
           for (let i = 0; i < checkboxes.length; i++) {
             checkboxes[i].checked = true;
           }
+
+          deliveryDatesContainer.classList.remove('invisible');
         } else {
           for (let i = 0; i < checkboxes.length; i++) {
             checkboxes[i].checked = false;
           }
+          deliveryDatesContainer.classList.add('invisible');
         }
       }
 
+      // СЛУШАТЕЛЬ СОБЫТИЯ НА ОТДЕЛЬНОМ ЧЕКБОКСЕ - переключает отображение отдельного товара в разделе доставки
       if (e.target.classList.contains('checkbox__input')) {
+        const id = +e.target.dataset['id'];
+        let elements = document.querySelectorAll(`.dlv[data-id="${id}"]`);
+
         if (!e.target.checked) {
           document.querySelector('.checkbox__input-all').checked = false;
+          elements.forEach(elem => elem.classList.add('invisible'));
+        } else {
+          document.querySelector('.checkbox__input-all').checked = true;
+          elements.forEach(elem => elem.classList.remove('invisible'));
         }
+
+        let deliveryContainers = document.querySelectorAll('.delivery__images');
+
+        deliveryContainers.forEach(elem => {
+          elements = elem.querySelectorAll('.dlv');
+          let visibility = [];
+          elements.forEach(el => {
+
+            if (!el.classList.contains('invisible')) {
+              visibility.push(true);
+            }
+
+            if (visibility.length === 0) {
+              elem.previousElementSibling.classList.add('invisible');
+            } else {
+              elem.previousElementSibling.classList.remove('invisible');
+            }
+          })
+
+        });
       }
     });
   }
